remove the 2GB limit on single malloc request, for rv64 system

+++ newlib/libc/stdlib/mallocr.c
@@ -2353,7 +2353,7 @@ Void_t* mALLOc(RARG bytes) RDECL size_t bytes;
   INTERNAL_SIZE_T nb  = request2size(bytes);  /* padded request size; */
 
   /* Check for overflow and just fail, if so. */
-  if (nb > INT_MAX || nb < bytes)
+  if (nb > LONG_MAX || nb < bytes)
   {
     RERRNO = ENOMEM;
     return 0;
@@ -2819,7 +2819,7 @@ Void_t* rEALLOc(RARG oldmem, bytes) RDECL Void_t* oldmem; size_t bytes;
   nb = request2size(bytes);
 
   /* Check for overflow and just fail, if so. */
-  if (nb > INT_MAX || nb < bytes)
+  if (nb > LONG_MAX || nb < bytes)
   {
     RERRNO = ENOMEM;
     return 0;
@@ -3055,7 +3055,7 @@ Void_t* mEMALIGn(RARG alignment, bytes) RDECL size_t alignment; size_t bytes;
   nb = request2size(bytes);
 
   /* Check for overflow. */
-  if (nb > INT_MAX || nb < bytes)
+  if (nb > LONG_MAX || nb < bytes)
   {
     RERRNO = ENOMEM;
     return 0;
